+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.5] - 2021-09-13
+### Changed
+- Replaced `value` slot `row`/`col` slot props with `labels` slot prop
+This new slot prop gathers `row`/`col` and adds field data, allowing cell value rendering based on field keys.
+
 ## [1.1.4] - 2021-08-20
 ### Changed
 - Replaced custom `HashTable` with `ManyKeysMap`